@@ -1,4 +1,4 @@
-// swift-tools-version: 5.10
+// swift-tools-version: 6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -18,6 +18,7 @@ let package = Package(
             name: "xxPROJECTxNAMExx"),
         .testTarget(
             name: "xxPROJECTxNAMExxTests",
-            dependencies: ["xxPROJECTxNAMExx"]),
+            dependencies: ["xxPROJECTxNAMExx"]
+        ),
     ]
 )
